requirements.txt
# Core dependencies for face recognition
face-recognition==1.3.0
opencv-python==4.12.0.88
numpy==1.26.4
Pillow==11.0.0

# Database
pymongo==4.15.5
python-dotenv==1.2.1

# API REST
Flask==3.1.2
flask-cors==6.0.2

# Reports generation
pandas==2.3.3
openpyxl==3.1.5

# Additional utilities
click==8.3.1
